vcpkg: add updateScript
This change allows you to update vcpkg by running this command:

  nix-shell --argstr package vcpkg maintainers/scripts/update.nix

The main motivation behind adding an update script is to make it easy to
know whether or not vcpkg can be updated automatically. There was at
least one occasion where we weren’t sure whether or not a change would
break automatic updates [1]. Now that we have an update script, we can
easily test whether or not automatic updates will work.

As an added bonus, this commit also allows us to automatically update
vcpkg without having to rely on nixpkgs-update.

[1]: <https://github.com/NixOS/nixpkgs/pull/329939#discussion_r1695155924>